Bump terraform-json to v0.22.1
There is an issue in v0.17.1 of terrraform-json where state files can't
be read when there are checks present, since checks are an arrray, and
the tool tries to marshal them to a strruct. This causes an issue with
the ShowStateFile function.

Bumping terraform-json to the latest version and running `go mod tidy`
to pull in its dependencies.

Original issue in tfexec:
https://github.com/hashicorp/terraform-exec/issues/415

Line which doesn't work currently in this repo:
https://github.com/opentofu/tofu-exec/blob/e5963f176ec5f21e5d059428189439290554c25c/tfexec/show.go#L95-L99

Patch to terraform-json:
https://github.com/hashicorp/terraform-json/pull/101

Changlog from v0.17.1 to v0.22.1:
https://github.com/hashicorp/terraform-json/releases